did the test stuff correctly this time hopefully
