+++ templates/list_all_movies.html
@@ -6,7 +6,4 @@
 <h1 class="mb-5">All Movies</h1>
 <p class="mb-3">See our list of movie ratings below</p>
 <!-- TODO: Feature 1 -->
-<p>{{movie_repository.get_movie_by_title("Jurassic Park").title}}</p>
-<p>{{movie_repository.get_movie_by_title("Jurassic Park").director}}</p>
-<p>{{movie_repository.get_movie_by_title("Jurassic Park").rating}}</p>
 {% endblock %}
