--- tests/unit/conftest.py
@@ -1,6 +1,0 @@
-import pytest
-from app import app
-
-@pytest.fixture(scope="module")
-def test_app():
-    return app.test_client()
