+++ tests/unit/test_create_movie.py
@@ -1,30 +1,18 @@
 # TODO: Feature 2
-def test_get_create_movie_page(test_app):
-    response = test_app.get('/movies/new')
-    
-    data = response.data.decode('utf-8')
-    
-    assert response.status_code == 200
-    assert '<h1 class="mb-5">Create Movie Rating</h1>' in data
-    assert '<p class="mb-3">Create a movie rating below</p>' in data
-    
+from app import movie_repository
+from src.repositories.movie_repository import get_movie_repository
 
-def test_create_movie_rating(test_app):
-    response = test_app.post("/movies", data={
-        "movie-title": "Jurassic Park",
-        "movie-director": "Steven Spielberg",
-        "movie-rating": "5"
-    }, follow_redirects=True)
+def test_no_movie_added_to_repository():
+    repo = get_movie_repository()
+    repo.clear_db()
+    assert len(repo.get_all_movies()) == 0
+
+def test_add_movie_to_repository():
+    movie_repository.create_movie('Jurassic Park', "Steven Spielberg", "5")
+    assert movie_repository.get_movie_by_title("Jurassic Park").title == "Jurassic Park"
+    assert movie_repository.get_movie_by_title("Jurassic Park").director == "Steven Spielberg"
+    assert movie_repository.get_movie_by_title("Jurassic Park").rating == "5"
+    assert type(movie_repository.get_movie_by_title("Jurassic Park").title) == type("")
+    assert type(movie_repository.get_movie_by_title("Jurassic Park").director) == type("")
+    assert type(movie_repository.get_movie_by_title("Jurassic Park").rating) == type("")
     
-    data = response.data.decode("utf-8")
-    
-    assert response.status_code == 200
-    assert '<p>Jurassic Park</p>' in data
-    assert '<p>Steven Spielberg</p>' in data
-    assert '<p>5</p>' in data
-    
-    
-def test_create_movie_rating_validation_error(test_app):
-    response = test_app.post('/movies', follow_redirects=True)
-    
-    assert response.status_code == 400
